split const and tokens from implementation
