+++ src/packages/language/collection/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_LANGUAGE_COLLECTION_ALIAS = 'Umb.Collection.Language';
