+++ src/packages/language/collection/index.ts
@@ -1,2 +1,2 @@
 export { UmbLanguageCollectionRepository } from './repository/index.js';
-export { UMB_LANGUAGE_COLLECTION_ALIAS } from './manifests.js';
+export { UMB_LANGUAGE_COLLECTION_ALIAS } from './constants.js';
