+++ src/packages/language/collection/manifests.ts
@@ -2,9 +2,8 @@ import { UMB_LANGUAGE_COLLECTION_REPOSITORY_ALIAS } from './repository/index.js'
 import { manifests as collectionRepositoryManifests } from './repository/manifests.js';
 import { manifests as collectionViewManifests } from './views/manifests.js';
 import { manifests as collectionActionManifests } from './action/manifests.js';
+import { UMB_LANGUAGE_COLLECTION_ALIAS } from './constants.js';
 import type { ManifestCollection, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_LANGUAGE_COLLECTION_ALIAS = 'Umb.Collection.Language';
 
 const collectionManifest: ManifestCollection = {
 	type: 'collection',
