+++ src/packages/language/collection/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_LANGUAGE_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.LanguageCollection';
