+++ src/packages/language/collection/repository/index.ts
@@ -1,2 +1,2 @@
-export { UMB_LANGUAGE_COLLECTION_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_LANGUAGE_COLLECTION_REPOSITORY_ALIAS } from './constants.js';
 export { UmbLanguageCollectionRepository } from './language-collection.repository.js';
