+++ src/packages/language/global-contexts/index.ts
@@ -1,1 +1,2 @@
 export * from './app-language.context.js';
+export { UMB_APP_LANGUAGE_CONTEXT } from './app-language.context-token.js';
