+++ src/packages/language/repository/detail/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_LANGUAGE_DETAIL_REPOSITORY_ALIAS = 'Umb.Repository.Language.Detail';
+export const UMB_LANGUAGE_DETAIL_STORE_ALIAS = 'Umb.Store.Language.Detail';
