+++ src/packages/language/repository/detail/index.ts
@@ -1,2 +1,2 @@
 export { UmbLanguageDetailRepository } from './language-detail.repository.js';
-export { UMB_LANGUAGE_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_LANGUAGE_DETAIL_REPOSITORY_ALIAS } from './constants.js';
