+++ src/packages/language/repository/detail/language-detail.repository.ts
@@ -1,6 +1,6 @@
 import type { UmbLanguageDetailModel } from '../../types.js';
 import { UmbLanguageServerDataSource } from './language-detail.server.data-source.js';
-import { UMB_LANGUAGE_DETAIL_STORE_CONTEXT } from './language-detail.store.js';
+import { UMB_LANGUAGE_DETAIL_STORE_CONTEXT } from './language-detail.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbDetailRepositoryBase } from '@umbraco-cms/backoffice/repository';
