+++ src/packages/language/repository/item/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_LANGUAGE_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.LanguageItem';
+export const UMB_LANGUAGE_STORE_ALIAS = 'Umb.Store.LanguageItem';
