+++ src/packages/language/repository/item/index.ts
@@ -1,3 +1,3 @@
 export { UmbLanguageItemRepository } from './language-item.repository.js';
-export { UMB_LANGUAGE_ITEM_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_LANGUAGE_ITEM_REPOSITORY_ALIAS } from './constants.js';
 export type { UmbLanguageItemModel } from './types.js';
