+++ src/packages/language/repository/item/language-item.repository.ts
@@ -1,5 +1,5 @@
 import { UmbLanguageItemServerDataSource } from './language-item.server.data-source.js';
-import { UMB_LANGUAGE_ITEM_STORE_CONTEXT } from './language-item.store.js';
+import { UMB_LANGUAGE_ITEM_STORE_CONTEXT } from './language-item.store.context-token.js';
 import type { UmbLanguageItemModel } from './types.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemRepositoryBase } from '@umbraco-cms/backoffice/repository';
