+++ src/packages/language/repository/item/manifests.ts
@@ -1,7 +1,5 @@
+import { UMB_LANGUAGE_ITEM_REPOSITORY_ALIAS, UMB_LANGUAGE_STORE_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestItemStore, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_LANGUAGE_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.LanguageItem';
-export const UMB_LANGUAGE_STORE_ALIAS = 'Umb.Store.LanguageItem';
 
 const itemRepository: ManifestRepository = {
 	type: 'repository',
